round: A Gimp/PS style capped opacity brush
The stroke is painted to it's own layer then blended
onto the target layer, which allows it to paint at
a fixed opacity.

+++ BaseKernels.cl
@@ -2,7 +2,7 @@ __kernel void circle(__global float4 *buf,
                               int     stride,
                               int     x,
                               int     y,
-                              int     r,
+                              float   r,
                               float4  color)
 {
   int gidx = get_global_id(0);
